rename crypto evtloop to edgeos_crypt and edgeos_evtloop
