+++ lib/crypto/edgeos_crypto.h
