+++ lib/crypto/keys/secpk256k1-key.pem
@@ -1,0 +1,5 @@
+-----BEGIN EC PRIVATE KEY-----
+MHQCAQEEIGf++w6nK6pyqoBhWwrEYzSK/iqVWmq/ZLNKn0w9XlvyoAcGBSuBBAAK
+oUQDQgAE6WdSXuRnBoBrX1X5WHTETKlEcxwM6KeD198FQELlfUQO8qF+I1AgTeir
+TeXUDmnkRjevikp4A3M3ugNYdIzhTQ==
+-----END EC PRIVATE KEY-----
