+++ lib/crypto/keys/secpk256k1-pubkey.pem
@@ -1,0 +1,4 @@
+-----BEGIN PUBLIC KEY-----
+MFYwEAYHKoZIzj0CAQYFK4EEAAoDQgAE6WdSXuRnBoBrX1X5WHTETKlEcxwM6KeD
+198FQELlfUQO8qF+I1AgTeirTeXUDmnkRjevikp4A3M3ugNYdIzhTQ==
+-----END PUBLIC KEY-----
