+++ lib/incl/edge_os.h
@@ -2,7 +2,8 @@
 #define __EDGE_OS_H__
 
 #include <edgeos_list.h>
-#include <evtloop.h>
+#include <edgeos_evtloop.h>
+#include <edgeos_crypto.h>
 #include <edgeos_netapi.h>
 #include <distcom.h>
 #include <fsapi.h>
@@ -10,8 +11,8 @@
 #include <prng.h>
 #include <edgeos_logger.h>
 #include <config_parser.h>
-#include <conv.h>
-#include <tokenparse.h>
+#include <edgeos_conv.h>
+#include <edgeos_tokparse.h>
 
 #ifdef OS_LINUX
 // QNX declares getopt() in unistd.h
