+++ lib/linux/net/ds/controller.c
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #include <getopt.h>
 #include <edgeos_list.h>
-#include <evtloop.h>
+#include <edgeos_evtloop.h>
 #include <edgeos_netapi.h>
 #include <dist_sdp.pb-c.h>
