+++ lib/util/edgeos_utils.h
