+++ lib/util/tokenparse.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 #include <string.h>
-#include <tokenparse.h>
+#include <edgeos_tokparse.h>
 
 int token_parser(const char *input, int input_len, char token, char *op, int op_len, int off)
 {
