+++ tests/test_evtloop.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <functional>
 extern "C" {
-#include <evtloop.h>
+#include <edgeos_evtloop.h>
 #include <edgeos_netapi.h>
 }
 #include <cstring>
