+++ tests/test_sysioctl.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 #include <stdint.h>
-#include <sysioctl.h>
+#include <edgeos_ioctl.h>
 
 int sysioctl_test(int argc, char **argv)
 {
@@ -31,7 +31,10 @@ int sysioctl_test(int argc, char **argv)
     if (ret == 0)
         printf("ip %s\n", ip);
 
+    ret = edge_os_get_hwaddr(NULL, NULL);
+
     uint8_t mac[6];
+
     ret = edge_os_get_hwaddr("enp67s0", mac);
     if (ret == 0)
         printf("mac : %02x:%02x:%02x:%02x:%02x:%02x\n", mac[0], mac[1], mac[2], mac[3], mac[4], mac[5]);
