+++ tests/transfer_test.c
@@ -6,7 +6,7 @@
 #include <sys/time.h>
 #include <getopt.h>
 #include <edgeos_sched.h>
-#include <evtloop.h>
+#include <edgeos_evtloop.h>
 #include <edgeos_netapi.h>
